Basket Analysis
Dashboard of a healthy products store to analyze which products sell best together with others. 
This kind of analysis are often performed by supermarkets and convenience stores,
but can also be done by restaurants(food industry), or local stores like in this case. 
For example, a branch of Walmart in USA performed this analysis and found out that beer was selling good along with diapers; 
with this they can target a specific segment or launch customized promotions.